requirements.txt
Django==3.2.4
django-crispy-forms==1.11.2
Faker==16.6.1
black==23.1.0